refactor(circles): Rename circle feature to circles
