+++ features/circles/utils.ts
@@ -1,12 +1,12 @@
 import JsonLdFactory from "@/utils/jsonLd";
-import { CircleSchema } from "@/services/circle";
+import { CircleSchema } from "@/services/circles";
 
 export function createCircleJsonLd(result: CircleSchema) {
   return JsonLdFactory.createEventBuilder()
-    .setId(`https://www.daoedu.tw/group/${result._id}`)
+    .setId(`https://www.daoedu.tw/circles/${result._id}`)
     .setName(result.title)
     .setDescription(result.content)
-    .setUrl(`https://www.daoedu.tw/group/${result._id}`)
+    .setUrl(`https://www.daoedu.tw/circles/${result._id}`)
     .setImage(result.photoURL)
     .setLocation(result.area)
     .setStartDate(result.createdDate)
